add caching for location slug checks
