+++ frontend/src/app/layout.tsx
@@ -15,6 +15,7 @@ import { GTMRouteTracker } from "@/contexts/GTMRouteTracker";
 import { Suspense } from "react";
 import { ConditionalClarity, ConditionalGA, ConditionalGTM, ConditionalGTMNoScript } from "@/components/analytics/Analytics";
 import { DEFAULT_CLARITY_ID, DEFAULT_GA_ID, DEFAULT_GTM_ID } from "@/lib/constants";
+import { prewarmLocationSlugCache } from "@/lib/location/locationSlugs";
 
 const alice = Alice({
   weight: ['400'],
@@ -55,6 +56,7 @@ export default function RootLayout({
 }: Readonly<{
   children: React.ReactNode;
 }>) {
+  prewarmLocationSlugCache();
   return (
     <html lang="en" className={alice.className}>
       <head>
